pyproject.toml
[project]
name = "nlp-hw"
version = "0.1.0"
description = "YouTube video summarizer (Russian) — local inference on Apple Silicon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp",
    "torch",
    "torchaudio",
    "pywhispercpp",
    "pyannote-audio",
    "langchain-openai",
    "python-dotenv",
    "rich",
]
